See if pytest is cool with moving these test fixtures into `utils`
